Add support for aspectRatioFixed, aspectRatioWidthFactor and aspectRatioHeightFactor attributes
aspectRatioWidthFactor and aspectRatioHeightFactor are used to determine the width or height of the layout prior to setup being completed (Setup consists of parsing the constrains and adding them to the solver)

aspectRatioWidthFactor: the width factor of the aspect ratio e.g 16
aspectRatioHeightFactor: the height factor of the aspect ratio e.g. 9

aspectRatioFixed: if set to true, the layouts height or width will be determined by the aspect ratio, not the constraints, before *and* after setup.